refactor(api): shared device DTOs in hypercolor-types::api
First domain of the shared API-contract extraction (doc 57 §5 Move 4):
device list/summary/zone/topology DTOs, Pagination, and the update and
identify request bodies now have one definition in
hypercolor-types::api, serialized by the daemon and deserialized by the
web UI and TUI — wire drift is a compile error instead of a runtime
parse failure.

The pairing vocabulary (DeviceAuthSummary, PairingDescriptor, flow
kinds, request/outcome types) moves from hypercolor-driver-api to
hypercolor-types::pairing (re-exported for drivers), because the web UI
had re-implemented all seven types by hand — it cannot depend on
driver-api. This also surfaces and fixes real drift the mirrors had
accumulated: led_count/total_leds were usize in the UI vs u32 on the
wire, and the UI's DeviceListResponse had silently dropped pagination.

+++ crates/hypercolor-ui/src/components/device_card.rs
@@ -348,7 +348,7 @@ pub fn DeviceCard(
         .map(|(i, z)| {
             (
                 topology_shape_svg(&z.topology),
-                z.led_count,
+                z.led_count as usize,
                 zone_palette[i % zone_palette.len()],
             )
         })
